Create a prometheus client
Part of the refactoring to organize the codebase. This moves the methods to
talk to prometheus into a class in its own module.

The unit tests have also been moved to a separate file.

closes https://github.com/CCI-MOC/openshift-usage-scripts/issues/69
