+++ .github/workflows/unit-tests.yaml
@@ -22,4 +22,4 @@ jobs:
 
       - name: Run unit tests
         run: |
-          python -m unittest openshift_metrics/tests/test_utils.py
+          python -m unittest openshift_metrics/tests/test_*
