+++ openshift_metrics/tests/test_utils.py
@@ -17,39 +17,6 @@ from unittest import TestCase, mock
 from openshift_metrics import utils
 import os
 
-class TestQueryMetric(TestCase):
-
-    @mock.patch('requests.Session.get')
-    @mock.patch('time.sleep')
-    def test_query_metric(self, mock_sleep, mock_get):
-        mock_response = mock.Mock(status_code=200)
-        mock_response.json.return_value = {"data": {
-            "result": "this is data"
-        }}
-        mock_get.return_value = mock_response
-
-        metrics = utils.query_metric('https://fake-url', 'fake-token', 'fake-metric', '2022-03-14', '2022-03-14')
-        self.assertEqual(metrics, "this is data")
-        self.assertEqual(mock_get.call_count, 1)
-
-    @mock.patch('requests.Session.get')
-    @mock.patch('time.sleep')
-    def test_query_metric_exception(self, mock_sleep, mock_get):
-        mock_get.return_value = mock.Mock(status_code=404)
-
-        self.assertRaises(Exception, utils.query_metric, 'https://fake-url', 'fake-token',
-                          'fake-metric', '2022-03-14', '2022-03-14')
-        self.assertEqual(mock_get.call_count, 3)
-
-    @mock.patch('requests.Session.get')
-    @mock.patch('time.sleep')
-    def test_query_metric_connection_error(self, mock_sleep, mock_get):
-        mock_get.side_effect = [ConnectionError]
-        self.assertRaises(ConnectionError, utils.query_metric, 'https://fake-url', 'fake-token',
-                  'fake-metric', '2022-03-14', '2022-03-14')
-        self.assertEqual(mock_get.call_count, 1)
-
-
 class TestGetNamespaceAnnotations(TestCase):
 
     @mock.patch('openshift_metrics.utils.requests.post')
